discount data
addDiscountToCategory: Validates request data, finds the specified category by ID, checks if a discount already exists, creates a new discount, and associates it with the category using a polymorphic relationship.

addDiscountToSubCategory: Validates request data, finds the specified subcategory by ID, checks if a discount already exists, creates a new discount, and associates it with the subcategory using a polymorphic relationship.

addDiscountToItem: Validates request data, finds the specified item by ID, checks if a discount already exists, creates a new discount, and associates it with the item using a polymorphic relationship.

getDiscounted: Validates the 'type' parameter, and depending on the type (1 for items, 2 for categories, 3 for subcategories), retrieves and returns the appropriate discounts in a JSON response.

update: Validates the incoming request data, finds the specified discount by ID, updates the discount with the validated data, and returns a success response.

destroy: Finds the specified discount by ID, deletes the discount, and returns a success response

Details are in the comments inside the controller